Start Poin
Here are the basic estructure of the game.
Assets: Some textures, a couple of models (lamp and a generator), I have build a really small maze like lab and I have implemented a FPSController in where I put a raycast coming from the camera to detect objects.
Finally a from the camera scrip we can call other functions for the actions of interactive objects, for example now I have an generator that powers up light  .