New interactive object + stuffs
Today I made a lot of things and is no good to all of this changes in one commit, sorry for that, I´ll keep that in mind in the future. Now whats new:
-Implemented a security door that has and open animation.
-Fixes on the way the objectives were changing when accomplishing and objetive, now spaming on and interactive object wont make the objectives change indefinilty.
- Now the function to advance in objetives is called were the interaction with the object that accomplish the objective occurs.
-A global event object and script were implemented to keep track of all the events that are ocurring in the game.

Note 1: All major clases work directly with the interaction script.
Note 2: Not sure whats better, to keep a script in each object that is wating to be detected by a raycast or simply identifying each objetct in the intreaction script, this is how is being doing right now. (probably option one is better, but for the sake of speed I wont change that except if a problem of optimizations is perceptible.